+++ anise/src/analysis/elements.rs
@@ -1,0 +1,17 @@
+/*
+ * ANISE Toolkit
+ * Copyright (C) 2021-onward Christopher Rabotin <christopher.rabotin@gmail.com> et al. (cf. AUTHORS.md)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * Documentation: https://nyxspace.com/
+ */
+
+/// Orbital element defines all of the supported orbital elements in ANISE, which are all built from a State.
+#[derive(Clone, Debug, PartialEq)]
+pub enum OrbitalElement {
+    SemiMajorAxis,
+    RAAN,
+    Eccentricity,
+}
